+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Please add your entries according to this format.
 
 ## Unreleased
 
+- Fix task caching for ktfmtCheckTask when project has multiple source sets (#288)
+
 ## Version 0.18.0 _(2024-04-13)_
 
 - Make `KtfmtCheckTask` cacheable